feat: 토큰에서 사용자 정보 가져오도록 수정

+++ src/main/java/makar/dev/common/security/dto/TokenDto.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class TokenDto {
-    private Integer userId;
+    private Long userId;
 
-    public static TokenDto of(Integer userId) {
+    public static TokenDto of(Long userId) {
         return new TokenDto(userId);
     }
 }
